refactoring absence components
