+++ muikku-frontend/muikkuV3/components/guider/dialogs/student.tsx
@@ -34,6 +34,7 @@ export type tabs =
   | "GUIDANCE_RELATIONS"
   | "STUDY_HISTORY"
   | "PEDAGOGICAL_SUPPORT"
+  | "ABSENCES"
   | "LANGUAGE_PROFILE";
 
 /**
@@ -67,11 +68,8 @@ interface StudentDialogProps {
  */
 const StudentDialog: React.FC<StudentDialogProps> = (props) => {
   const { isOpen, student, onClose } = props;
-
   const { guider } = useSelector((state: StateType) => state);
-
   const dispatch = useDispatch();
-
   const { t } = useTranslation(["common"]);
 
   /** Number of contact logs to display per page */
@@ -156,6 +154,12 @@ const StudentDialog: React.FC<StudentDialogProps> = (props) => {
       name: t("labels.languageProfile"),
       type: "guider-student",
       component: <LanguageProfile userId={student.basic?.userEntityId} />,
+    },
+    {
+      id: "ABSENCES",
+      name: t("labels.absences"),
+      type: "guider-student",
+      component: <Absences userId={student.basic?.userEntityId} />,
     },
   ];
